requirements.txt
loguru
openai==1.6.1
ipykernel==6.17.1
tqdm
python-dotenv
chromadb==0.4.21
datasets==2.16.0
requests
matplotlib
seaborn
transformers
dirtyjson
scikit-learn
tiktoken
git+https://github.com/lapp0/outlines.git@fix-vllm-group-generation
openpyxl
#
# Local models only:
#
# torch
# accelerate==0.25.0
# bitsandbytes==0.41.3
# git+https://github.com/viktor-ferenczi/outlines.git
# pydantic==2.0.0
# vllm==0.2.7